// angular/src/app/api/model/clienteFisicoListResponse.ts
/**
 * OpenAPI definition
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Dados resumidos de cliente pessoa física para listas
 */
export interface ClienteFisicoListResponse {
  /**
   * ID do cliente
   */
  id?: number;
  /**
   * Nome completo
   */
  nome?: string;
  /**
   * CPF (formato: 000.000.000-00)
   */
  cpf?: string;
  /**
   * E-mail do cliente
   */
  email?: string;
  /**
   * Status do cliente
   */
  estaAtivo?: boolean;
}